requirements.txt
arrow==1.2.3
asgiref==3.5.2
attrs==22.1.0
binaryornot==0.4.4
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cookiecutter==2.1.1
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-discover-runner==1.0
django-role-permissions==3.1.1
exceptiongroup==1.0.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typed-ast==1.5.4
types-cryptography==3.3.23
tzdata==2022.4
urllib3==1.26.11
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
